+++ octopod/vision/dataset.py
@@ -64,9 +64,9 @@ class OctopodImageDataset(Dataset):
         return len(self.x)
 
 
-class TonksImageDatasetMultiLabel(TonksImageDataset):
+class OctopodImageDatasetMultiLabel(OctopodImageDataset):
     """
-    Subclass of TonksImageDataset used for multi-label tasks
+    Subclass of OctopodImageDataset used for multi-label tasks
 
     Parameters
     ----------
@@ -77,12 +77,12 @@ class TonksImageDatasetMultiLabel(TonksImageDataset):
         classes in the multi-label task. For a 4 class multi-label task
         a sample list would be [1,0,0,1]
     transform: str or list of PyTorch transforms
-        specifies how to preprocess the full image for a Tonks image model
-        To use the built-in Tonks image transforms, use the strings: `train` or `val`
+        specifies how to preprocess the full image for a Octopod image model
+        To use the built-in Octopod image transforms, use the strings: `train` or `val`
         To use custom transformations supply a list of PyTorch transforms
     crop_transform: str or list of PyTorch transforms
-        specifies how to preprocess the center cropped image for a Tonks image model
-        To use the built-in Tonks image transforms, use strings `train` or `val`
+        specifies how to preprocess the center cropped image for a Octopod image model
+        To use the built-in Octopod image transforms, use strings `train` or `val`
         To use custom transformations supply a list of PyTorch transforms
     """