+++ tests/fixtures/learner_fixtures.py
@@ -3,7 +3,7 @@ import pytest
 from torch.utils.data import DataLoader
 
 from octopod import MultiDatasetLoader
-from octopod.vision.dataset import TonksImageDataset
+from octopod.vision.dataset import OctopodImageDataset
 
 
 @pytest.fixture()
@@ -67,14 +67,14 @@ def test_train_val_loaders():
     df = pd.DataFrame({'image_loc': ['sample_data/tonks.jpeg'],
                        'category': [1]})
 
-    task1_dataset = TonksImageDataset(
+    task1_dataset = OctopodImageDataset(
         x=df['image_loc'],
         y=df['category'],
         transform='train',
         crop_transform='train'
     )
 
-    task2_datset = TonksImageDataset(
+    task2_datset = OctopodImageDataset(
         x=df['image_loc'],
         y=df['category'],
         transform='train',