+++ tonks/__init__.py
@@ -1,8 +1,0 @@
-from ._version import __version__
-
-from octopod.dataloader import MultiDatasetLoader
-from octopod.ensemble import *
-from octopod.learner import MultiTaskLearner, MultiInputMultiTaskLearner
-from octopod.learner_utils import *
-from octopod.text import *
-from octopod.vision import *