+++ tonks/text/__init__.py
@@ -1,2 +1,0 @@
-from octopod.text.dataset import TonksTextDataset, TonksTextDatasetMultiLabel
-from octopod.text.models import BertForMultiTaskClassification