+++ tonks/vision/__init__.py
@@ -1,2 +1,0 @@
-from octopod.vision.dataset import TonksImageDataset, TonksImageDatasetMultiLabel
-from octopod.vision.models import ResnetForMultiTaskClassification